Added comments and refactored main.c

+++ Breakout/draw.h
@@ -1,5 +1,6 @@
 #pragma once
 
+/* Define color macros for easy use */
 #define RED 1.0f, 0.0f, 0.0f
 #define GREEN 0.0f, 1.0f, 0.0f
 #define BLUE 0.0f, 0.0f, 1.0f
@@ -8,9 +9,10 @@
 #define WHITE 1.0f, 1.0f, 1.0f
 #define BLACK 0.0f, 0.0f, 0.0f
 
+/* Define PI for use by drawCirc */
 #define PI 3.14159265f
 
-void drawRect(float x, float y, float width, float height);
-void drawCirc(float x, float y, float radius);
+void drawRect(int x, int y, int width, int height);
+void drawCirc(int x, int y, int radius);
 void setColor(float r, float g, float b);
 void setBackgroundColor(float r, float g, float b);
